Added docblock '''templates''' to each group of defines with a @link reference to the file(s) using the defines in this group. This should provide at least one-way linkage for each define '''to''' the source file using them. Please test - if this works, the // start-of-line comments for each group may be deleted.
To accomplish two-way linking, the only solution would be a @uses tag for each constant used by a file or function/method.


Also added a parameter replacement comment back: phpDocumentor docblocks are intended for programmers; translators will tend to work just with the single file and will need consistent documentation '''in the file'''  to provide context and usage information.

#340