refs #498
A new parameter was added to the Format() method, it is named $format_option.
It is a semicolon separated list of strings, and actually, the only parameter defined is 'page', which means we are formatting an entire page, and it involves that closing tags and generating header ids should be done.

The two handlers handlers/show/show.php and handlers/edit/edit.php were just slightly edited to add this new parameter. (Closing tags and generating id are done when previewing in edit.php)

Removed the core parameter named formatter_recursion, which become obsolete.