refs #97
By default, RewriteEngine should be set to Off to ensure that no problem will be encountered before install. If it is set to On and the user tries to install on a subdirectory, the installer may be not reachable (The server issues a 500 error code if the RewriteBase is not correct or absent).