COORD of [940]
This commit is just about consistency and does not affect how Wikka works. Search Form (TextSearch) at the bottom of the page should have URL referring to page TextSearch, no matter which page are currently viewed. Before this commit, it refers to the same page that is viewed. I said It does not affect how Wikka works because this value is not used since there is an <input> tag inside the form.

refs #670, #575